Trait reth_rpc_api::servers::EngineApiServer

pub trait EngineApiServer<Engine: EngineTypes>: Sized + Send + Sync + 'static {

    // Required methods
    fn new_payload_v1<'life0, 'async_trait>(
        &'life0 self,
        payload: ExecutionPayloadV1
    ) -> Pin<Box<dyn Future<Output = RpcResult<PayloadStatus>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn new_payload_v2<'life0, 'async_trait>(
        &'life0 self,
        payload: ExecutionPayloadInputV2
    ) -> Pin<Box<dyn Future<Output = RpcResult<PayloadStatus>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn new_payload_v3<'life0, 'async_trait>(
        &'life0 self,
        payload: ExecutionPayloadV3,
        versioned_hashes: Vec<B256>,
        parent_beacon_block_root: B256
    ) -> Pin<Box<dyn Future<Output = RpcResult<PayloadStatus>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn fork_choice_updated_v1<'life0, 'async_trait>(
        &'life0 self,
        fork_choice_state: ForkchoiceState,
        payload_attributes: Option<Engine::PayloadAttributes>
    ) -> Pin<Box<dyn Future<Output = RpcResult<ForkchoiceUpdated>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn fork_choice_updated_v2<'life0, 'async_trait>(
        &'life0 self,
        fork_choice_state: ForkchoiceState,
        payload_attributes: Option<Engine::PayloadAttributes>
    ) -> Pin<Box<dyn Future<Output = RpcResult<ForkchoiceUpdated>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn fork_choice_updated_v3<'life0, 'async_trait>(
        &'life0 self,
        fork_choice_state: ForkchoiceState,
        payload_attributes: Option<Engine::PayloadAttributes>
    ) -> Pin<Box<dyn Future<Output = RpcResult<ForkchoiceUpdated>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_payload_v1<'life0, 'async_trait>(
        &'life0 self,
        payload_id: PayloadId
    ) -> Pin<Box<dyn Future<Output = RpcResult<ExecutionPayloadV1>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_payload_v2<'life0, 'async_trait>(
        &'life0 self,
        payload_id: PayloadId
    ) -> Pin<Box<dyn Future<Output = RpcResult<ExecutionPayloadEnvelopeV2>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_payload_v3<'life0, 'async_trait>(
        &'life0 self,
        payload_id: PayloadId
    ) -> Pin<Box<dyn Future<Output = RpcResult<ExecutionPayloadEnvelopeV3>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_payload_bodies_by_hash_v1<'life0, 'async_trait>(
        &'life0 self,
        block_hashes: Vec<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<ExecutionPayloadBodiesV1>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_payload_bodies_by_range_v1<'life0, 'async_trait>(
        &'life0 self,
        start: U64,
        count: U64
    ) -> Pin<Box<dyn Future<Output = RpcResult<ExecutionPayloadBodiesV1>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn exchange_transition_configuration<'life0, 'async_trait>(
        &'life0 self,
        transition_configuration: TransitionConfiguration
    ) -> Pin<Box<dyn Future<Output = RpcResult<TransitionConfiguration>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn exchange_capabilities<'life0, 'async_trait>(
        &'life0 self,
        capabilities: Vec<String>
    ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<String>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

    // Provided method
    fn into_rpc(self) -> RpcModule<Self>
       where Engine: Send + Sync + 'static + DeserializeOwned { ... }

}

Server trait implementation for the EngineApi RPC API.

Required Methods

fn new_payload_v1<'life0, 'async_trait>( &'life0 self, payload: ExecutionPayloadV1 ) -> Pin<Box<dyn Future<Output = RpcResult<PayloadStatus>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

See also https://github.com/ethereum/execution-apis/blob/6709c2a795b707202e93c4f2867fa0bf2640a84f/src/engine/paris.md#engine_newpayloadv1 Caution: This should not accept the withdrawals field

fn new_payload_v2<'life0, 'async_trait>( &'life0 self, payload: ExecutionPayloadInputV2 ) -> Pin<Box<dyn Future<Output = RpcResult<PayloadStatus>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

See also https://github.com/ethereum/execution-apis/blob/584905270d8ad665718058060267061ecfd79ca5/src/engine/shanghai.md#engine_newpayloadv2

fn new_payload_v3<'life0, 'async_trait>( &'life0 self, payload: ExecutionPayloadV3, versioned_hashes: Vec<B256>, parent_beacon_block_root: B256 ) -> Pin<Box<dyn Future<Output = RpcResult<PayloadStatus>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Post Cancun payload handler

See also https://github.com/ethereum/execution-apis/blob/main/src/engine/cancun.md#engine_newpayloadv3

fn fork_choice_updated_v1<'life0, 'async_trait>( &'life0 self, fork_choice_state: ForkchoiceState, payload_attributes: Option<Engine::PayloadAttributes> ) -> Pin<Box<dyn Future<Output = RpcResult<ForkchoiceUpdated>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

See also https://github.com/ethereum/execution-apis/blob/6709c2a795b707202e93c4f2867fa0bf2640a84f/src/engine/paris.md#engine_forkchoiceupdatedv1

Caution: This should not accept the withdrawals field in the payload attributes.

fn fork_choice_updated_v2<'life0, 'async_trait>( &'life0 self, fork_choice_state: ForkchoiceState, payload_attributes: Option<Engine::PayloadAttributes> ) -> Pin<Box<dyn Future<Output = RpcResult<ForkchoiceUpdated>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Post Shanghai forkchoice update handler

This is the same as forkchoiceUpdatedV1, but expects an additional withdrawals field in the payloadAttributes, if payload attributes are provided.

See also https://github.com/ethereum/execution-apis/blob/6709c2a795b707202e93c4f2867fa0bf2640a84f/src/engine/shanghai.md#engine_forkchoiceupdatedv2

Caution: This should not accept the parentBeaconBlockRoot field in the payload attributes.

fn fork_choice_updated_v3<'life0, 'async_trait>( &'life0 self, fork_choice_state: ForkchoiceState, payload_attributes: Option<Engine::PayloadAttributes> ) -> Pin<Box<dyn Future<Output = RpcResult<ForkchoiceUpdated>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Post Cancun forkchoice update handler

This is the same as forkchoiceUpdatedV2, but expects an additional parentBeaconBlockRoot field in the the payloadAttributes, if payload attributes are provided.

See also https://github.com/ethereum/execution-apis/blob/main/src/engine/cancun.md#engine_forkchoiceupdatedv3

fn get_payload_v1<'life0, 'async_trait>( &'life0 self, payload_id: PayloadId ) -> Pin<Box<dyn Future<Output = RpcResult<ExecutionPayloadV1>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

See also https://github.com/ethereum/execution-apis/blob/6709c2a795b707202e93c4f2867fa0bf2640a84f/src/engine/paris.md#engine_getpayloadv1

Returns the most recent version of the payload that is available in the corresponding payload build process at the time of receiving this call.

Caution: This should not return the withdrawals field

Note:

Provider software MAY stop the corresponding build process after serving this call.

fn get_payload_v2<'life0, 'async_trait>( &'life0 self, payload_id: PayloadId ) -> Pin<Box<dyn Future<Output = RpcResult<ExecutionPayloadEnvelopeV2>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

See also https://github.com/ethereum/execution-apis/blob/6709c2a795b707202e93c4f2867fa0bf2640a84f/src/engine/shanghai.md#engine_getpayloadv2

Returns the most recent version of the payload that is available in the corresponding payload build process at the time of receiving this call. Note:

Provider software MAY stop the corresponding build process after serving this call.

fn get_payload_v3<'life0, 'async_trait>( &'life0 self, payload_id: PayloadId ) -> Pin<Box<dyn Future<Output = RpcResult<ExecutionPayloadEnvelopeV3>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Post Cancun payload handler which also returns a blobs bundle.

See also https://github.com/ethereum/execution-apis/blob/main/src/engine/cancun.md#engine_getpayloadv3

Returns the most recent version of the payload that is available in the corresponding payload build process at the time of receiving this call. Note:

Provider software MAY stop the corresponding build process after serving this call.

fn get_payload_bodies_by_hash_v1<'life0, 'async_trait>( &'life0 self, block_hashes: Vec<BlockHash> ) -> Pin<Box<dyn Future<Output = RpcResult<ExecutionPayloadBodiesV1>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

See also https://github.com/ethereum/execution-apis/blob/6452a6b194d7db269bf1dbd087a267251d3cc7f8/src/engine/shanghai.md#engine_getpayloadbodiesbyhashv1

fn get_payload_bodies_by_range_v1<'life0, 'async_trait>( &'life0 self, start: U64, count: U64 ) -> Pin<Box<dyn Future<Output = RpcResult<ExecutionPayloadBodiesV1>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

See also https://github.com/ethereum/execution-apis/blob/6452a6b194d7db269bf1dbd087a267251d3cc7f8/src/engine/shanghai.md#engine_getpayloadbodiesbyrangev1

Returns the execution payload bodies by the range starting at start, containing count blocks.

WARNING: This method is associated with the BeaconBlocksByRange message in the consensus layer p2p specification, meaning the input should be treated as untrusted or potentially adversarial.

Implementors should take care when acting on the input to this method, specifically ensuring that the range is limited properly, and that the range boundaries are computed correctly and without panics.

fn exchange_transition_configuration<'life0, 'async_trait>( &'life0 self, transition_configuration: TransitionConfiguration ) -> Pin<Box<dyn Future<Output = RpcResult<TransitionConfiguration>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

See also https://github.com/ethereum/execution-apis/blob/6709c2a795b707202e93c4f2867fa0bf2640a84f/src/engine/paris.md#engine_exchangetransitionconfigurationv1

Note: This method will be deprecated after the cancun hardfork:

Consensus and execution layer clients MAY remove support of this method after Cancun. If no longer supported, this method MUST be removed from the engine_exchangeCapabilities request or response list depending on whether it is consensus or execution layer client.

fn exchange_capabilities<'life0, 'async_trait>( &'life0 self, capabilities: Vec<String> ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<String>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

See also https://github.com/ethereum/execution-apis/blob/6452a6b194d7db269bf1dbd087a267251d3cc7f8/src/engine/common.md#capabilities

Provided Methods

fn into_rpc(self) -> RpcModule<Self>
where Engine: Send + Sync + 'static + DeserializeOwned,

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.

Object Safety

This trait is not object safe.