Trait reth_rpc_api::servers::OtterscanServer

pub trait OtterscanServer: Sized + Send + Sync + 'static {

    // Required methods
    fn has_code<'life0, 'async_trait>(
        &'life0 self,
        address: Address,
        block_number: Option<BlockId>
    ) -> Pin<Box<dyn Future<Output = RpcResult<bool>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_api_level<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = RpcResult<u64>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_internal_operations<'life0, 'async_trait>(
        &'life0 self,
        tx_hash: TxHash
    ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<InternalOperation>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_transaction_error<'life0, 'async_trait>(
        &'life0 self,
        tx_hash: TxHash
    ) -> Pin<Box<dyn Future<Output = RpcResult<String>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn trace_transaction<'life0, 'async_trait>(
        &'life0 self,
        tx_hash: TxHash
    ) -> Pin<Box<dyn Future<Output = RpcResult<TraceEntry>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_block_details<'life0, 'async_trait>(
        &'life0 self,
        block_number: BlockNumberOrTag
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<BlockDetails>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_block_details_by_hash<'life0, 'async_trait>(
        &'life0 self,
        block_hash: B256
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<BlockDetails>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_block_transactions<'life0, 'async_trait>(
        &'life0 self,
        block_number: BlockNumberOrTag,
        page_number: usize,
        page_size: usize
    ) -> Pin<Box<dyn Future<Output = RpcResult<OtsBlockTransactions>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn search_transactions_before<'life0, 'async_trait>(
        &'life0 self,
        address: Address,
        block_number: BlockNumberOrTag,
        page_size: usize
    ) -> Pin<Box<dyn Future<Output = RpcResult<TransactionsWithReceipts>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn search_transactions_after<'life0, 'async_trait>(
        &'life0 self,
        address: Address,
        block_number: BlockNumberOrTag,
        page_size: usize
    ) -> Pin<Box<dyn Future<Output = RpcResult<TransactionsWithReceipts>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_transaction_by_sender_and_nonce<'life0, 'async_trait>(
        &'life0 self,
        sender: Address,
        nonce: u64
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<Transaction>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_contract_creator<'life0, 'async_trait>(
        &'life0 self,
        address: Address
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<ContractCreator>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

    // Provided method
    fn into_rpc(self) -> RpcModule<Self> { ... }

}

Server trait implementation for the Otterscan RPC API.

Required Methods

fn has_code<'life0, 'async_trait>( &'life0 self, address: Address, block_number: Option<BlockId> ) -> Pin<Box<dyn Future<Output = RpcResult<bool>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Check if a certain address contains a deployed code.

fn get_api_level<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = RpcResult<u64>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Very simple API versioning scheme. Every time we add a new capability, the number is incremented. This allows for Otterscan to check if the node contains all API it needs.

fn get_internal_operations<'life0, 'async_trait>( &'life0 self, tx_hash: TxHash ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<InternalOperation>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return the internal ETH transfers inside a transaction.

fn get_transaction_error<'life0, 'async_trait>( &'life0 self, tx_hash: TxHash ) -> Pin<Box<dyn Future<Output = RpcResult<String>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Given a transaction hash, returns its raw revert reason.

fn trace_transaction<'life0, 'async_trait>( &'life0 self, tx_hash: TxHash ) -> Pin<Box<dyn Future<Output = RpcResult<TraceEntry>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Extract all variations of calls, contract creation and self-destructs and returns a call tree.

fn get_block_details<'life0, 'async_trait>( &'life0 self, block_number: BlockNumberOrTag ) -> Pin<Box<dyn Future<Output = RpcResult<Option<BlockDetails>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Tailor-made and expanded version of eth_getBlockByNumber for block details page in Otterscan.

fn get_block_details_by_hash<'life0, 'async_trait>( &'life0 self, block_hash: B256 ) -> Pin<Box<dyn Future<Output = RpcResult<Option<BlockDetails>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Tailor-made and expanded version of eth_getBlockByHash for block details page in Otterscan.

fn get_block_transactions<'life0, 'async_trait>( &'life0 self, block_number: BlockNumberOrTag, page_number: usize, page_size: usize ) -> Pin<Box<dyn Future<Output = RpcResult<OtsBlockTransactions>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get paginated transactions for a certain block. Also remove some verbose fields like logs.

fn search_transactions_before<'life0, 'async_trait>( &'life0 self, address: Address, block_number: BlockNumberOrTag, page_size: usize ) -> Pin<Box<dyn Future<Output = RpcResult<TransactionsWithReceipts>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Gets paginated inbound/outbound transaction calls for a certain address.

fn search_transactions_after<'life0, 'async_trait>( &'life0 self, address: Address, block_number: BlockNumberOrTag, page_size: usize ) -> Pin<Box<dyn Future<Output = RpcResult<TransactionsWithReceipts>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Gets paginated inbound/outbound transaction calls for a certain address.

fn get_transaction_by_sender_and_nonce<'life0, 'async_trait>( &'life0 self, sender: Address, nonce: u64 ) -> Pin<Box<dyn Future<Output = RpcResult<Option<Transaction>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Gets the transaction hash for a certain sender address, given its nonce.

fn get_contract_creator<'life0, 'async_trait>( &'life0 self, address: Address ) -> Pin<Box<dyn Future<Output = RpcResult<Option<ContractCreator>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Gets the transaction hash and the address who created a contract.

Provided Methods

fn into_rpc(self) -> RpcModule<Self>

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.

Object Safety

This trait is not object safe.